✨feat : 명함 상세 타입  추가
